fix formpublicaciones

+++ lib/src/componentes/publicaciones/cubit/formpublicaciones_cubit.dart
@@ -10,19 +10,21 @@ import 'package:uuid/uuid.dart';
 part 'formpublicaciones_state.dart';
 
 class FormPublicacionesCubit extends Cubit<FormPublicacionesState> {
-  
   FormPublicacionesCubit() : super(FormPublicacionesState.initial());
 
-
-  getDataPublicacionUpdate(Publicacion publicacion,List<Empresa> empresas){
-   
-    final index = empresas.indexWhere((empresa) => publicacion.empresa.id == empresa.id);
-    emit(state.copyWith(
-         empresa  : publicacion.empresa,
-         selecionada: index,
-         imagenes : publicacion.imagenes.map((e) => ImageFile(nombre: e,isaFile: false)).toList()
-    ));
+  getDataPublicacionUpdate(Publicacion? publicacion, List<Empresa> empresas) {
+    if (publicacion != null) {
+      final index = empresas
+          .indexWhere((empresa) => publicacion.empresa.id == empresa.id);
+      emit(state.copyWith(
+          empresa: publicacion.empresa,
+          selecionada: index,
+          imagenes: publicacion.imagenes
+              .map((e) => ImageFile(nombre: e, isaFile: false))
+              .toList()));
+    }
   }
+
   selectEmpresa(Empresa empresa, int index) =>
       emit(state.copyWith(empresa: empresa, selecionada: index));
 
@@ -32,9 +34,10 @@ class FormPublicacionesCubit extends Cubit<FormPublicacionesState> {
   addImagen(File file) {
     emit(state.copyWith(
         imagenes: List.of(state.imagenes)
-          ..add(ImageFile(nombre: '${Uuid().v4()}.jpg', file: file, isaFile: true))));
+          ..add(ImageFile(
+              nombre: '${Uuid().v4()}.jpg', file: file, isaFile: true))));
   }
-  
+
   updateImagen(File file, ImageFile? imagen, int index) async {
     emit(state.copyWith(
         imagenes: List.of(state.imagenes)
@@ -46,16 +49,13 @@ class FormPublicacionesCubit extends Cubit<FormPublicacionesState> {
                   file: file,
                   isaFile: true))));
     if (imagen!.isaFile) {
-      if(await imagen.file!.exists())
-      imagen.file!.delete();
+      if (await imagen.file!.exists()) imagen.file!.delete();
     }
   }
 
   Future deleteFiles() async {
-    state.imagenes.forEach((imagen) async { 
-       await imagen.file!.delete();
+    state.imagenes.forEach((imagen) async {
+      await imagen.file!.delete();
     });
   }
-
-  
 }
